Generalized the registries component, can now just simply add to the objectTypes in objectTypes.js
